requirements.txt
accelerate==0.29.2
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.42.0
anyio==4.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bip-utils==2.9.3
bitarray==2.9.2
boto3==1.36.16
botocore==1.36.16
browser-use==0.1.36
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.0
cachetools==4.2.4
cbor2==5.6.4
cdp==0.0.2
cdp-agentkit-core==0.0.11
cdp-langchain==0.0.13
cdp-sdk==0.15.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.4.24
ckzg==2.0.1
click==8.0.1
coincurve==20.0.0
coloredlogs==15.0.1
courlan==1.3.2
cprint==1.2.2
crcmod==1.7
cryptography==44.0.0
cssselect==1.2.0
cytoolz==0.12.3
dataclasses-json==0.6.4
datasets==2.20.0
dateparser==1.2.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
discord.py==2.4.0
distro==1.9.0
ecdsa==0.19.0
ed25519-blake2b==1.4.1
elevenlabs==1.2.2
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
executing==2.0.1
faiss-cpu==1.10.0
fastapi==0.110.0
fastchat==0.1.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.3
filetype==1.2.0
firebase-admin==5.0.0
fireworks-ai==0.15.12
Flask==2.2.2
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.14.0
google-cloud-translate==2.0.1
google-crc32c==1.5.0
google-generativeai==0.8.4
google-resumable-media==2.7.0
googleapis-common-protos==1.67.0rc1
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
gunicorn==20.1.0
h11==0.14.0
hexbytes==1.2.1
html2text==2024.2.26
htmldate==1.9.3
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.2
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jusText==3.0.1
kubernetes==29.0.0
langchain==0.3.14
langchain-anthropic==0.3.3
langchain-aws==0.2.12
langchain-chroma==0.1.0
langchain-community==0.3.13
langchain-core==0.3.34
langchain-experimental==0.0.59
langchain-fireworks==0.2.7
langchain-google-genai==2.0.8
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-pinecone==0.0.3
langchain-text-splitters==0.3.4
langchainhub==0.1.15
langgraph==0.2.60
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.4
lmnr==0.4.59
lru-dict==1.2.0
lxml==5.3.0
lxml_html_clean==0.4.1
MainContentExtractor==0.0.4
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.1
marshmallow==3.21.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
mypy==1.14.1
mypy-extensions==1.0.0
neo4j==5.22.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.17.3
openai==1.61.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-langchain==0.38.4
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-instrumentation-sqlalchemy==0.51b0
opentelemetry-instrumentation-threading==0.51b0
opentelemetry-instrumentation-urllib3==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-semantic-conventions-ai==0.4.2
opentelemetry-util-http==0.51b0
orjson==3.10.12
outcome==1.3.0.post0
overrides==7.7.0
p5==0.8.4
packaging==23.2
pandas==2.2.2
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pillow_heif==0.17.0
pinecone-client==3.2.2
playwright==1.50.0
posthog==3.11.0
prompt-toolkit==3.0.43
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
py-sr25519-bindings==0.2.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.6
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.17.2
PyJWT==2.10.1
PyNaCl==1.5.0
PyOpenGL-accelerate==3.1.9
pyp5js==0.7.3
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.6
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==6.1.2
pytube==15.0.0
pytz==2025.1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.35.0
regex==2023.12.25
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rlp==4.0.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.11.2
safetensors==0.4.3
scandir==1.10.0
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.28.1
sentence-transformers==3.4.1
sentencepiece==0.2.0
setuptools==75.8.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
skia-python==87.6
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
sse-starlette==2.1.0
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tavily-python==0.5.0
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tinysegmenter==0.3
tld==0.13
tldextract==5.1.3
tokenizers==0.19.1
toolz==0.12.1
torch==2.2.2
tqdm==4.66.4
trafilatura==2.0.0
traitlets==5.14.3
Transcrypt==3.9.0
transformers==4.42.3
trio==0.28.0
trio-websocket==0.11.1
tweepy==4.14.0
typer==0.12.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
uv==0.5.29
uvicorn==0.29.0
uvloop==0.19.0
watchdog==2.1.9
watchfiles==0.21.0
wcwidth==0.2.13
web3==7.8.0
web3-input-decoder==0.1.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.18.1
